QDeviceDiscovery: Add static discovery support for new-style multitouch devices.
These don't report ABS_X/ABS_Y, but are still a touchscreen. This is commonly
seen on Android hardware.

As a bonus, we clean up the flow control through device discovery: by using
memset, we can ignore the error value of ioctl() and just check the bits we
need.

We can also close() the FD instantly after running the ioctl calls.